requirements.txt
alembic==1.16.5
blinker==1.9.0
catboost==1.2.8
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
et_xmlfile==2.0.0
Flask==3.1.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.60.1
graphviz==0.21
greenlet==3.2.4
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.7
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.9.4
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
Pillow==10.1.0
plotly==5.18.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.44
tenacity==9.1.2
threadpoolctl==3.6.0
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
WTForms==3.2.1
xgboost==2.1.4
xlrd==2.0.2
zipp==3.23.0
